refactor: move QuotaPlanChangedEvent to core package for cross-plugin access
Move event definitions from internal/event to core/ so other plugins
can subscribe to quota plan changes. Update quota service imports to
use pluginCore instead of internal event package.
